requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
argparse==1.4.0
asgiref==3.6.0; python_full_version >= '3.7.0'
certifi==2022.12.7; python_version >= '3.6'
cffi==1.15.1; platform_python_implementation == 'CPython' and sys_platform == 'win32'
charset-normalizer==3.1.0; python_full_version >= '3.7.0'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cors==1.0.1
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django==4.1.7
djangorestframework==3.14.0
filelock==3.10.7; python_full_version >= '3.7.0'
future==0.18.3; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gevent==22.10.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
greenlet==2.0.2; platform_python_implementation == 'CPython'
idna==3.4; python_version >= '3.5'
pillow==9.4.0
pycparser==2.21
pysocks==1.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
requests-file==1.5.1
requests==2.28.2; python_version < '4' and python_full_version >= '3.7.0'
setuptools==67.6.1; python_full_version >= '3.7.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3; python_version >= '3.5'
tldextract==3.4.0; python_full_version >= '3.7.0'
tzdata==2023.3; sys_platform == 'win32'
urllib3==1.26.15; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
zope.event==4.6
zope.interface==6.0; python_full_version >= '3.7.0'
